+++ packages/common/src/TableId.ts
@@ -1,12 +1,12 @@
 import { Hex, stringToHex, hexToString, sliceHex, concatHex } from "viem";
 
 export class TableId {
-  namespace: string;
-  name: string;
+  readonly namespace: string;
+  readonly name: string;
 
   constructor(namespace: string, name: string) {
-    this.namespace = namespace;
-    this.name = name;
+    this.namespace = namespace.substring(0, 16);
+    this.name = name.substring(0, 16);
   }
 
   toString(): string {
@@ -18,7 +18,10 @@ export class TableId {
   }
 
   static toHex(namespace: string, name: string): Hex {
-    return concatHex([stringToHex(namespace, { size: 16 }), stringToHex(name, { size: 16 })]);
+    return concatHex([
+      stringToHex(namespace.substring(0, 16), { size: 16 }),
+      stringToHex(name.substring(0, 16), { size: 16 }),
+    ]);
   }
 
   static fromHex(hex: Hex): TableId {