+++ packages/common/src/utils/bigIntSort.ts
@@ -1,0 +1,3 @@
+export function bigIntSort(a: bigint, b: bigint): -1 | 0 | 1 {
+  return a < b ? -1 : a > b ? 1 : 0;
+}