+++ packages/store-sync/src/sqlite/schemaVersion.ts
@@ -1,0 +1,3 @@
+// When this is incremented, it forces all indexers to reindex from scratch the next time they start up.
+// Only use this when the schemas change, until we get proper schema migrations.
+export const schemaVersion = 1;