+++ packages/solhint-config-mud/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Change Log
 
+## 2.0.0-next.0
+
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.